refactor: "resolveJsonModule": false
