--- src/anthem.json
@@ -1,1 +1,0 @@
-["동","해","물","과","백","두","산","이","마","르","고","닳","도","록","하","느","님","이","보","우","하","사","우","리","나","라","만","세","무","궁","화","삼","천","리","화","려","강","산","대","한","사","람","대","한","으","로","길","이","보","전","하","세","남","산","위","에","저","소","나","무","철","갑","을","두","른","듯","바","람","서","리","불","변","함","은","우","리","기","상","일","세","무","궁","화","삼","천","리","화","려","강","산","대","한","사","람","대","한","으","로","길","이","보","전","하","세","가","을","하","늘","공","활","한","데","높","고","구","름","없","이","밝","은","달","은","우","리","가","슴","일","편","단","심","일","세","무","궁","화","삼","천","리","화","려","강","산","대","한","사","람","대","한","으","로","길","이","보","전","하","세","이","기","상","과","이","맘","으","로","충","성","을","다","하","여","괴","로","우","나","즐","거","우","나","나","라","사","랑","하","세","무","궁","화","삼","천","리","화","려","강","산","대","한","사","람","대","한","으","로","길","이","보","전","하","세"]
